Refactor Filter classes into a separate filter folder, along side very minor styling refactoring
