+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/counting/tweet/TopSecondDegreeByCountRequestForTweet.java
@@ -21,7 +21,7 @@ import java.util.Map;
 import java.util.Set;
 
 import com.twitter.graphjet.algorithms.RecommendationType;
-import com.twitter.graphjet.algorithms.ResultFilterChain;
+import com.twitter.graphjet.algorithms.filters.ResultFilterChain;
 import com.twitter.graphjet.algorithms.counting.TopSecondDegreeByCountRequest;
 
 import it.unimi.dsi.fastutil.longs.Long2DoubleMap;
