+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/salsa/SalsaRequest.java
@@ -18,7 +18,7 @@
 package com.twitter.graphjet.algorithms.salsa;
 
 import com.twitter.graphjet.algorithms.RecommendationRequest;
-import com.twitter.graphjet.algorithms.ResultFilterChain;
+import com.twitter.graphjet.algorithms.filters.ResultFilterChain;
 import com.twitter.graphjet.hashing.SmallArrayBasedLongToDoubleMap;
 
 import it.unimi.dsi.fastutil.longs.Long2DoubleMap;
