+++ graphjet-core/src/test/java/com/twitter/graphjet/algorithms/salsa/SalsaSelectResultsTest.java
@@ -30,9 +30,9 @@ import static org.junit.Assert.assertEquals;
 import com.twitter.graphjet.algorithms.BipartiteGraphTestHelper;
 import com.twitter.graphjet.algorithms.ConnectingUsersWithMetadata;
 import com.twitter.graphjet.algorithms.RecommendationInfo;
-import com.twitter.graphjet.algorithms.RequestedSetFilter;
-import com.twitter.graphjet.algorithms.ResultFilter;
-import com.twitter.graphjet.algorithms.ResultFilterChain;
+import com.twitter.graphjet.algorithms.filters.RequestedSetFilter;
+import com.twitter.graphjet.algorithms.filters.ResultFilter;
+import com.twitter.graphjet.algorithms.filters.ResultFilterChain;
 import com.twitter.graphjet.algorithms.counting.tweet.TweetRecommendationInfo;
 import com.twitter.graphjet.algorithms.salsa.fullgraph.SalsaInternalState;
 import com.twitter.graphjet.bipartite.api.BipartiteGraph;
